# app/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для управления пользовательскими подписками: создание, получение, обновление, удаление, список с пагинацией и подсчёт суммы по периодам.",
        "title": "API Подписок",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/addSub": {
            "post": {
                "description": "Создаёт новую подписку для указанного пользователя и сервиса.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Создать подписку",
                "parameters": [
                    {
                        "description": "Тело запроса для создания подписки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/deleteSub": {
            "delete": {
                "description": "Удаляет подписку по UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Удалить подписку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID подписки (UUID)",
                        "name": "sub_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/getSub": {
            "get": {
                "description": "Возвращает информацию о подписке по UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Получить подписку по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID подписки (UUID)",
                        "name": "sub_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/listSubs": {
            "post": {
                "description": "Возвращает список подписок, отфильтрованных по сервису, пользователю и диапазону дат. Поддерживает курсорную пагинацию.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Список подписок",
                "parameters": [
                    {
                        "description": "Фильтры и курсор пагинации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/totalSubs": {
            "post": {
                "description": "Рассчитывает суммарную стоимость подписок по заданным фильтрам за пересекающийся период.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Общая сумма по подпискам",
                "parameters": [
                    {
                        "description": "Фильтры для расчёта суммы",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/updateSub": {
            "put": {
                "description": "Обновляет ежемесячную стоимость и дату окончания подписки.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Обновить подписку",
                "parameters": [
                    {
                        "description": "Тело запроса для обновления подписки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.AddReq": {
            "type": "object",
            "properties": {
                "monthly_fee": {
                    "description": "ежемесячная стоимость",
                    "type": "integer",
                    "example": 499
                },
                "num_months": {
                    "description": "длительность подписки в месяцах",
                    "type": "integer",
                    "example": 12
                },
                "service_name": {
                    "description": "название сервиса",
                    "type": "string",
                    "example": "Netflix"
                },
                "start_date": {
                    "description": "дата начала (MM_YYYY)",
                    "type": "string",
                    "example": "01-2025"
                },
                "user_id": {
                    "description": "ID пользователя (UUID)",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "handlers.CursorIn": {
            "type": "object",
            "properties": {
                "last_id": {
                    "description": "последний ID (UUID)",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "last_start": {
                    "description": "последняя дата (MM-YYYY)",
                    "type": "string",
                    "example": "01-2025"
                },
                "limit": {
                    "description": "размер страницы (по умолчанию 50)",
                    "type": "integer",
                    "example": 50
                }
            }
        },
        "handlers.ListReq": {
            "type": "object",
            "properties": {
                "cursor": {
                    "description": "курсор пагинации (необязательно)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handlers.CursorIn"
                        }
                    ]
                },
                "end_date": {
                    "description": "конечная дата периода (необязательно)",
                    "type": "string",
                    "example": ""
                },
                "service_name": {
                    "description": "фильтр по названию сервиса (необязательно)",
                    "type": "string",
                    "example": "Netflix"
                },
                "start_date": {
                    "description": "начальная дата периода (необязательно)",
                    "type": "string",
                    "example": "01-2025"
                },
                "user_id": {
                    "description": "фильтр по пользователю (UUID, необязательно)",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "handlers.UpdateReq": {
            "type": "object",
            "properties": {
                "end_date": {
                    "description": "новая дата окончания (MM-YYYY)",
                    "type": "string",
                    "example": "01-2025"
                },
                "monthly_fee": {
                    "description": "новая ежемесячная стоимость",
                    "type": "integer",
                    "example": 599
                },
                "sub_id": {
                    "description": "ID подписки (UUID)",
                    "type": "string",
                    "example": "2f1e4e2a-1a7d-4e6b-a222-3cb3f92f0a11"
                }
            }
        },
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP код ответа",
                    "type": "integer"
                },
                "data": {
                    "description": "Данные ответа (опционально)"
                },
                "message": {
                    "description": "Сообщение для пользователя",
                    "type": "string"
                }
            }
        }
    },
    "tags": [
        {
            "description": "Операции с подписками пользователей: создание, получение, обновление, удаление, список с пагинацией и подсчёт суммарной стоимости.",
            "name": "Subscriptions"
        }
    ]
}